fix: wrap netlify functions in directory to be able to zip static json mapping file as well
