+++ .gitignore
@@ -50,4 +50,4 @@ Thumbs.db
 
 # Generated files for dynamic route pre rendering
 routes.txt
-src/functions/utils/case-sensitive-video-id-mapping.json
+src/functions/videos/case-sensitive-video-id-mapping.json
