+++ src/functions/utils/youtube-api.js
@@ -1,31 +1,29 @@
-const fetch = require("node-fetch");
-const {getCaseSensitiveYoutubeVideoId }= require('./case-sensitive-video-id-mapper');
+const fetch = require('node-fetch');
 const apiKey = process.env.GOOGLE_API_KEY;
-const apiBase = (path) =>
+const apiBase = path =>
   `https://www.googleapis.com/youtube/v3/${path}?key=${apiKey}`;
 
-const getPlaylist = async (playlistId) => {
+const getPlaylist = async playlistId => {
   return await (
     await fetch(
       `${apiBase(
-        "playlistItems"
+        'playlistItems'
       )}&part=snippet&maxResults=50&playlistId=${playlistId}`
     )
   ).json();
 };
 
-const getVideos = async (videoId) => {
-  const caseSensitiveYoutubeVideoId = getCaseSensitiveYoutubeVideoId(videoId)
+const getVideos = async videoId => {
   return await (
     await fetch(
       `${apiBase(
-        "videos"
-      )}&part=snippet%2Cstatistics&id=${caseSensitiveYoutubeVideoId}&maxResults=50`
+        'videos'
+      )}&part=snippet%2Cstatistics&id=${videoId}&maxResults=50`
     )
   ).json();
 };
 
 module.exports = {
   getPlaylist,
-  getVideos,
+  getVideos
 };
